feat: Third-party block support (#22)
* build: Add `@wordpress/dependency-extraction-webpack-plugin`

Source which modules should be externalized.

* fix: Manually delete Jetpack plugin header causing CORS errors

This custom header causes CORS errors. Although settings the mode to
'cors' should prevent this header, incorrect middleware order results in
setting the header.

* fix: Prevent error when fetching a full URL

`@wordpress/api-fetch` supports fetching both a `path` and `url`. When
passing the former, we need to guard against an undefined `path.`

* feat: Add remote editor entry

Enable fetching remote editor assets from a site.

* feat: iOS loads remote editor

* fix: Prevent erroneously cleared content

It appears a race condition results in the content briefly displaying
before these lines reset out the content entirely.

* refactor: Improve `setupEditor` parameter type adherence

The second parameter (edits) appears to expect and object. The third
parameter is optional.

* style: Inline small object keys and values

Align with core code and improve readability.

* feat: Conditionally load remote editor

Rely upon plugin-provided remote editor for simplify proof-of-concept
testing.

* fix: API namespace utility supports optional preceding slash

It is possible to fetch relative paths, we should support this to avoid
cryptic errors.

* fix: Precede editor assets path with slash

Ensure the correct path is fetched.

* fix: Remote editor endpoint relies upon site URL

This plugin-provided endpoint cannot be retrieved via WPCOM REST API
URLs. We must rely upon the core REST API path.

* build: Capture remote editor build output

* build: Mitigate cache invalidation issues

Without unique hashes, it is near impossible to invalidate cached
assets.

* build: Capture remote editor build output

* fix: Remove erroneous, duplicative forward slash

A prior change to support paths without a preceding forward slash
introduced a bug where an extra forward slash was inserted into the
middle of the string replacement.

* fix: Reinstate missing trailing forward slash

Without this, only the first of the target two path segments are
captured.

* fix: Add missing trailing forward slash

* build: Capture remote editor build output

* build: Simplify Vite scripts

It was unnecessary to have multiple commands to build or preview
production builds.

* refactor: Remove seemingly unnecessary base configuration

The default is `/`, which should suffice.

* build: Capture latest build output

* build: Add remote editor dev server script

* fix: Remove duplicative entity addition

This already occurs in core logic. This line duplicated the logic and
resulted in unexpected clearing of editor content.

* fix: Reinstate base location configuration

The default of `/` incorrectly sources from the root of the project,
rather than the nested build directory.

* build: Capture latest build output

* feat: Conditionally load remote editor with feature flag

* fix: Ensure correct block editor registry is utilized

Without disabling the sub-registry, a separate, duplicative registry was
utilized, which caused certain block selectors to return empty results.

* build: Capture latest build output

* refactor: Leverage local `remote.html` file

Previously, a remote copy of `remote.html` existing on the targeted site
was utilized to avoid CORS errors when requesting the editor assets
endpoint. However, those CORS errors do not appear to occur now, which
may mean they were unrelated.