fix: Stabilize Android media uploads (#29)
* build: Synchronize AGP version

Address errors originating from version conflicts with the
WordPress-Android project.

* build: Synchronize Gradle version

Address version incompatibility by mirroring the WordPress-Android
project.

* fix: Reinstate `GBKit` global injection

It is not possible to use `evaluateJavascript` within the
`shouldInterceptRequest` callback, as it occurs to early. This resulted
in the absence of the `GBKit` global.

* refactor: Avoid seemingly unnecessary nullable value

Defaulting this value to `false` should suffice.

* fix: Prevent arbitrarily discard weak references

The weak references resulted in unexpectedly null callback values, which
meant that the media upload callback would occasionally fail.

* Revert "fix: Reinstate `GBKit` global injection"

This reverts commit f75562316b90ce4aea8065d587f84729f3068021.

* fix: Reinstate `GBKit` global injection

It is not possible to use `evaluateJavascript` within the
`shouldInterceptRequest` callback, as it occurs to early. This resulted
in the absence of the `GBKit` global.

* refactor: Address lint warnings

* feat: Catch editor configuration JSON parse errors

Mitigate cryptic errors from unexpected editor configuration values.