fix: Replace unsupported editor architecture
This largely reverts
https://github.com/wordpress-mobile/GutenbergKit/pull/24, which resulted
in several regressions. Namely, oddities in the block toolbar emptying
and empty Patterns/Media tabs in the block inserter.

The previous changes was implemented to workaround an experimental API
that is no longer exported after
https://github.com/WordPress/gutenberg/pull/64892. That remains an
issue, but the new workaround is to patch Gutenberg itself to export
this necessary API with `patch-package`.