fix: Replace unsupported editor architecture (#50)
* build: Add patch-package

* build: Configure patch-package

* fix: Replace unsupported editor architecture

This largely reverts
https://github.com/wordpress-mobile/GutenbergKit/pull/24, which resulted
in several regressions. Namely, oddities in the block toolbar emptying
and empty Patterns/Media tabs in the block inserter.

The previous changes was implemented to workaround an experimental API
that is no longer exported after
https://github.com/WordPress/gutenberg/pull/64892. That remains an
issue, but the new workaround is to patch Gutenberg itself to export
this necessary API with `patch-package`.

* fix: Disable interception `blob:` URLs leading to an Android crash

While the reason for why the `blob:` requests trigger is unknown, they
originate when utilizing the block inserter's Patterns tab. The requests
seemingly originate from the embedded editor previews rendered for each
pattern.