refactor: Hoist editor logic (#59)
* fix: Reinstate editor post title ref

This ref was mistakenly removed in the form of combining it with the ref
that tracks the _content_ of the post title. The reinstated ref is used
for auto-focusing the UI element itself.

* refactor: Relocate bridge logic to hook

Reduce the scope of concerns for the visual editor component.

* refactor: Relocate editor setup to hook

Reduce the scope of concerns for the visual editor component.

* refactor: Relocate editor ready check

Separate ready check from settings configuration.

* refactor: Relocate GBKit settings logic to hook

Reduce the scope of concerns for the visual editor component.

* refactor: Hoist editor logic into editor component

Prepare for utilizing this logic outside of the visual editor--namely,
within the text editor.

* refactor: Hoist editor provider to editor

Prepare for text editor.

* refactor: Hoist snackbars as editor child

Align code structure with the Gutenberg project.

* refactor: Relocate `useEditorStyles` to the visual editor component

This hook is used solely within this component at this time.

* refactor: Relocate `useMediaUpload` to the editor component

This hook is used solely within this component at this time.