fix: Avoid blurring the editor
Blurring the editor was done to ensure the latest text composition was
persisted to the DOM before the host app requested the latest value.
However, blurring the editor results in dismissing the virtual keyboard.
We can instead trigger a `compositionend` event to persist the latest
values without impacting the user experience.