refactor: Simplify content change bridge listener (#72)
* refactor: Simplify content change bridge listener

This is simplified for two reasons:

1. The JavaScript implementation does send the title and content values,
   so the method signature is misleading.
2. The host app is expected to retrieve the title and content whenever
   it decides to persist the latest changes. The `onContentChange` event
   is intended to notify the host of new content.

This aligns GutenbergKit with the approach used for the Gutenberg Mobile
and Aztec editors.

* fix: Avoid blurring the editor

Blurring the editor was done to ensure the latest text composition was
persisted to the DOM before the host app requested the latest value.
However, blurring the editor results in dismissing the virtual keyboard.
We can instead trigger a `compositionend` event to persist the latest
values without impacting the user experience.

* refactor: Remove disabled ESLint rule

It is preferred to use an explicit element, rather than the `document`.