fix: Improve autosave and unsaved changes detection (#104)
* fix: Improve autosave and unsaved changes detection

Previously, host apps determined whether content changes were present by
comparing the content received from GutenbergKit to the latest
server-provided post content. This is problematic as the server makes
slight modifications to content markup, creating perpetually changed
content.

To address this, we need GutenbergKit to communicate changed content.
We accomplish by sending a new `changed` parameter alongside the title
and content. Rather than updating the post and content refs on every
content change, we now do so at the time at which the host app requests
the latest content. This ensures the host and GutenbergKit synchronize
their latest perception of changes.

We also now rely upon `AutosaveMonitor` to throttle how often we request
the host app persist the latest changes, rather than a generic store
subscription.

* task: Capture build output

* fix: Copy latest block registration changes from local editor

We must register blocks prior to attempts to serialize the raw post
content within the editor.

This change was not tested. It should be tested once we re-engage work
on the remote editor.