fix: Open user-initiated link navigation in OS browser
The URL checking was added to ensure tapping links within the editor
opened in the OS browser, rather than the WebView. However, basing this
upon the URL is problematic, as JavaScript can also trigger navigation
events. Instead, we now open the URL in the browser if it is
user-initiated--i.e., a user tapped a link.