refactor: Avoid private useBlockEditorSettings
The `useBlockEditorSettings` function is not exported, even as a private
API. Using it is brittle, at best, and the approach is not possible when
sourcing WordPress package scripts from remote sites.

These changes still patch the `@wordpress/block-editor` package to (1)
allow the "quick inserter" to open the primary inserter and (2) enable
the Media tab in the inserter. Long-term, we should remove these patches
and replace them with changes to core.