refactor: Host provides editor settings (#114)
* feat: Limit block types to an allowed list

To improve stability, the remote, site-specific editor un-registers
block types that are not on the allow list.

* fix: Seed GBKit global from local storage

This is largely in place for development experience purposes, not user
facing. It allows the editor WebView to reload and retains its current
GBKit global.

* fix: Guard against undefined allowed block types configuration

Avoid exceptions invoking array methods on `undefined`.

* Revert "fix: Seed GBKit global from local storage"

This reverts commit 38fe2c7728b79668ebcd904605011bbacc5938f2.

* hoist settings

* feat: Expose bridge method for updating editor settings

* Revert "feat: Expose bridge method for updating editor settings"

This reverts commit 2e1f1464ac65899dc3a5b23af12b59978a34dee9.

* feat: Improve editor start sequence

* refactor: Separate warm up mode load behaviour

To allow the host app to manage editor start sequence, we do not start
within `viewDidLoad`.

* refactor: Remote editor relies upon host app settings

* refactor: Android receives editor settings from host app

Hoist editor settings fetching and configuration to the host app to
allow more control of the editor initialization flow.

* refactor: Pass Android editor settings as JSON string

Avoid difficulties typing complex JSON structures.

* refactor: Clean up logic

Remove unnecessary or unused code.

* task: Capture build output

* task: Capture build output

* Make `isWarmupMode` immutable (#117)

---------

Co-authored-by: Jeremy Massel <1123407+jkmassel@users.noreply.github.com>