fix: Workaround circular dependency
A circular dependency causes the dynamic import to never resolve. Using
a Promise circumvents the issue, as it appears to function differently
than top-level await.

Ideally, we resolve the circular dependency. However, using Rollup's
`manualChunks` configuration does not appear to be an option, as it
results in preloading of `@wordpress` modules. The preloading results in
untranslated strings as `@wordpress` modules load before the globals
they rely upon for localization, which are set by `setLocaleData`.