refactor: Remove unnecessary dependency injection
Because of the latest dynamic import strategy and Vite managing
externalizing `@wordpress` modules for the remote editor, we no longer
need to use DI for these functions.

For the local editor, the imported modules will be bundled and imported.
For the remote editor, the imported modules will be replaced with global
`window.wp` references.