fix: Mitigate Android text composition oddities
Forcibly ending text composition led to odd outcomes--e.g., erroneously
dismissing the inline inserter, resetting editor history stack.

It was originally added to ensure the latest text in the editor is
captured before providing it to the host app for persistence. This
appears necessary for Samsung device keyboards only. For the time being,
text may be lost on these devices if the user does not first blur the
text input before saving or exiting the editor.