fix: Scrolling along edges breaks text editing (#135)
Existing Gutenberg `preventFocusCapture` logic relies upon `pointerdown`
and `pointerup` events for temporarily disabling all text blocks
`contenteditable` status. This is down to mitigate issues with block
selection for `flex` elements.

However, `pointerup` events are not always triggered on touch devices
when a touch turns into a swipe. When you scroll with your finger
outside of the block canvas, the `pointerup` callback is never invoked.
Scrolling twice or more leads to a stale `value` in the callback, and a
state where `contenteditable` is perpetually disabled.