fix: Embed renders (#140)
* fix: Ensure WebView handles embed requests

The previous allow list approach meant one-off requests were often sent
to the OS browser. E.g., a YouTube embed launched the OS browser rather
than displaying the embedded video within the editor.

* fix: Add body class required for responsive embeds

This class is typically added by WordPress core when a theme enables the
`responsive-embeds` feature. Enabling it seems like a sounds default for
the mobile editor. Otherwise, content will not resize and overflow.

https://developer.wordpress.org/block-editor/how-to-guides/themes/theme-support/#responsive-embedded-content

* fix: Correctly size embeds

This is host specific code that should be hoisted to the host app.

* build: Ensure internal `@wordpress/url` import is not externalized

The `src/utils/api-fetch.js` utility now imports `@wordpress/url`. To
prevent Vite from attempting to externalize that import, we must check
if the `module` (i.e., the matched import syntax) is one of the modules
that should not be externalized.