feat: Mark content outside of popovers inert (#145)
* feat: Modalize the block inserter and inspector

Improve accessibility by hiding content outside of the modals from
screen readers.

* fix: Prevent nested inspector popovers dismissing the block inspector

Opening a nested settings--e.g., Text color--presents a popover and
moves focus to the new popover. In that case, we do not want to dismiss
the block inspector housing the button that opened the popover.

* fix: Correctly size global button text

Buttons rendered within popovers reside in the root body element, thus
we need to hoist this global style override higher.

* style: Remove unused import, fix typo

* docs: Document `Inserter` popover props patch

Allow setting popover props for the `Inserter` component, so we can
improve the mobile screen reader experience by marking it as a modal
dialog.