fix: Localize remote editor
The remote editor relies upon accessing `@wordpress` modules via the
`window.wp` global. Most of these references are handled by
externalization performed by Vite. However, the `@wordpress/i18n` module
is a dependency of `@wordpress/api-fetch`, which we do not want to
externalize. So, the `import` of `@wordpress/i18n` is not correctly
externalized for the remote editor.

As a workaround, we check for and utilize the global if it is available.
Long term, we should align the local and remote editor architecture so
that they load in the same manner. At which point, we could remove the
conditional and always reference the global.