fix: Localize remote editor (#150)
* fix: Localize remote editor

The remote editor relies upon accessing `@wordpress` modules via the
`window.wp` global. Most of these references are handled by
externalization performed by Vite. However, the `@wordpress/i18n` module
is a dependency of `@wordpress/api-fetch`, which we do not want to
externalize. So, the `import` of `@wordpress/i18n` is not correctly
externalized for the remote editor.

As a workaround, we check for and utilize the global if it is available.
Long term, we should align the local and remote editor architecture so
that they load in the same manner. At which point, we could remove the
conditional and always reference the global.

* fix: Mitigate cryptic ES module failures

Occasionally, the local editor fails to load entirely. Sometimes it
display no errors, other times it'll throw exceptions related to
references `process`. Guarding the `process` reference will help prevent
the `process` error in particular.

* fix: Remote editor localization timeout

Dynamic importing the localization resulted in a timeout due to a cyclic
dependency with the remote entry file. Specifically, with vendor assets
such as Vite helper utilities.

* refactor: Replace anonymous functions with function declarations

Improve readability of top-level logic.

* refactor: Replace compat function with Vite import rewrite

Avoid boilerplate code by relying upon Vite to replace the import with a
global reference. The global `wp.i18n` reference is defined by the
scripts loaded from the site.

* docs: Note importance of promises for dynamic editor loads

* refactor: Local editor utilizes promises rather than async/await

Align implementation with the remote editor, which uses promises to
avoid timeouts from circular dependencies.

* refactor: Rename remote editor functions for consistency

Align with local editor naming.