feat: Bypass native bridge timeout via `dev_mode` query parameter
Simplify testing the editor in a web browser through disabling the error
displayed when the native bridge configuration is missing. The error is
useful in the context of a native host app running the editor, but
prevents easy testing of the editor web app outside of a native host
app.