refactor: Improve debugging (#164)
* style: Format CLAUDE.md

* feat: Bypass native bridge timeout via `dev_mode` query parameter

Simplify testing the editor in a web browser through disabling the error
displayed when the native bridge configuration is missing. The error is
useful in the context of a native host app running the editor, but
prevents easy testing of the editor web app outside of a native host
app.

* refactor: Extract shared bridge event dispatcher utility

* refactor: Log bridge events for debugging

* refactor: Allow setting logger level with URL query parameters

Simplify debugging in browser environments.

* feat: Disable automatic redirects in dev mode

Automatic redirects from the remote editor to the local editor make
debugging errors difficult.

* docs: Instruct Claude on project logging standards

* refactor: Minimize dev mode bypass logic

Lowering the bypass means more of the existing logic occurs, decreasing
the likelihood the bypass results in unexpected behavior.

* docs: Expand documentation