fix: API requests utilize single auth mechanism (#166)
* fix: Avoid cookies disrupting token authentication

GutenbergKit is intended to authentication REST API requests via
authentication tokens. If WordPress authentication cookies are present
in the request, it will take precedent over the authentication token
present in the request. This can result in request authentication
failures.

* test: Tests default to `jsdom` environment

Simplify test set up by defaulting to the most common environment for
this project.

* test: Assert API fetch utility authentication middleware

* refactor: Rename middleware and remove currying

* docs: Document `tokenAuthMiddleware`

* build: Upgrade caniuse database

Avoid out-of-date warnings logged in `npm` commands.