fix: Disable Android WebView preloading
This optimization logic leads to the editor UI disappearing due to race
conditions in the host WordPress app. Once we resolve the race
conditions, we should reinstate this preloading.

The race condition involves the WebView loading twice. Once for
preloading, and again when GutenbergKit is "started" with editor
settings. Between the first and second, the editor visibility is set to
invisible. The second load does not set the visibilility to visible, as
the `didFireEditorLoaded` guard logic only allows this to occur once.