fix: Replace WebView pool with warmup utilities
the benefits of pooling WebViews became moot once we stopped loading a
URL during the initialization sequence. It now provides no benefit.

Additionally, the pooled WebViews conflicted with the new warmup
utilities. It was possible for an editor launch to reuse an in-progress
warmup WebView. Once the warmup destroyed its WebView, the editor began
failing.