fix: Prevent multiple Android WebView starts (#175)
* chore: Ignore Kotlin build directory

* fix: Disable rudimentary WebView preloading

This load did not consider the editor configuration, meaning it often
loaded the incorrect URL.

Additionally, because the editor reveal logic currently only runs on
first load, this preload intermittently led to an invisible editor;
where the first load revealed the editor and the second, configured load
hid the editor.

1. Preload
2. Hide editor
3. Load URL
4. Reveal editor
5. Start editor with configuration
6. Hide editor
7. Load URL

* feat: Expand Android warmup to load all editor assets

Load and cache not only the manifest, but the editor assets themselves.

* refactor: De-duplicate Android editor load listeners

It appears we added multiple listeners for the same event at some point.
Also, we did not clean up one of the two listeners during tear down.

* refactor: Extracted magic number to named constant

* fix: Replace WebView pool with warmup utilities

the benefits of pooling WebViews became moot once we stopped loading a
URL during the initialization sequence. It now provides no benefit.

Additionally, the pooled WebViews conflicted with the new warmup
utilities. It was possible for an editor launch to reuse an in-progress
warmup WebView. Once the warmup destroyed its WebView, the editor began
failing.