Add Configuration Builder (and tests) (#146)
* Add EditorConfigurationBuilder

* Add EditorConfigurationBuilderTests

* Make Test Suite run on macOS

* docs: Document all editor configuration options

* test: Expand builder tests

* test: Fix EditorManifestTests for macOS runs

Apply iOS conditionals used in other, existing code.

* ci: Revert Swift test script device configuration removal

This was changed while making the tests pass when running on macOS, but
it is unclear why the original CI script was changed. Presumably we
should run the tests with a fixed device on the CI.

https://github.com/wordpress-mobile/GutenbergKit/pull/146/commits/debc1c74d5d460ac59363edf5117026e1f3ffe01#diff-76ed074a9305c04054cdebb9e9aad2d818052b07091de1f20cad0bbac34ffb52L62

* test: Remove builder cookie assertions

This functionality was removed in the trunk branch.

---------

Co-authored-by: David Calhoun <github@davidcalhoun.me>