fix: Block customization without editor settings (#180)
* feat: Set default colors and gradients for sites without settings

If a site does not have the Gutenberg plugin enabled, the editor
settings endpoint is unavailable. We set a default color palette and
gradients to enable customizing block styles.

* fix: Include preset editor styles

Ensures preset styles are loaded--e.g., utility classes for colors.

* chore: Update `useEditorStyles` copy with latest from Gutenberg core

Bring our copy of this private utility up to date.

* refactor: Utilize `useEditorSetting` added styles parameter

Embrace new abilities of this private function copy.

* fix: Default colors render for separators

Include default color utility classes. This is typically added by
WordPress' admin via PHP. The manual addition is necessary for sites
without the GutenbergKit plugin, which do not provide editor settings
and theme styles.

* fix: Apply default styles when editor settings provide none

Even if the theme style preferences is enabled, we must apply the
default styles if:

1. The theme provides no style via a block.json;
2. Or the site lacks the experimental Gutenberg REST API endpoint
   providing editor settings and styles.

* docs: Further document nuanced editor settings fall back logic