fix: Force visual editor mode to ensure consistency with native host UI
GutenbergKit provides no way to configure the initial editor mode. When
GutenbergKit persists a different mode in its local storage, this leads
to a mismatch in native UI and the editor mode. Default to visual avoids
any mismatches.

Additionally, forcing visual mode on initialization circumvents the fact
that the current logic does not trigger the `onEditorLoaded` event when
the text editor loads. This led to the native host never receiving the
event. The result was an indefinite loading indicator until the user
toggled the editor mode.