refactor: Remove remote editor variant (#210)
* task: Remove remote editor variant logic

The default editor now supports plugins, so the variant is not longer
necessary.

* refactor: Rename `bundled-editor` to `editor-environment`

With the removal of the remote editor variant, the "bundled" term no
longer communicates intent. The utility manages setting up the editor
environment with a specific sequence.

* test: Assert editor environment

* refactor: Remove unused editor loader filtering

* docs: Remove unnecessary entry point reference

* refactor: Remove unused `allowedBlockTypes` parameter

This is now handled in the `editor-environment` logic as it depends upon
the plugins assets fetching.

* test: Update tests to match latest logic

* task: Remove lingering remote editor variant references

* task: Clarify iOS demo app note

* task: Expand Android demo app documentation

* task: Relocate Android demo source note to bottom

* task: Restore original bundled editor note

* task: Relocate remote editors note

* task: Ensure Android demo app receives correct state

The dynamic Android demo app note regarding whether `make build` or
`make dev-server` is powering the editor was not working before.

* task: Android demo app optimizes keyboard for URL input