task: Update project configuration, scripts, and documentation (#229)
* docs: Fix README typo

* task: Improve Makefile command speed

Avoid reinstalling dependencies and fetching translations unless there
is an explicit need to do so. This improve repetitive command speed.

* task: Halt release script if localization string fetching fails

We should not publish the release if there were failures fetching the
l10n strings.

* docs: Ensure each script is documented for Claude

* task: Capture more details when fetching localization strings fails

* task: Retry localization string fetches twice

* task: Batch locale string fetching requests

Improve script speed.

* task: Ensure CI fetches locale strings

* task: Ensure command invoked directly always run

It's confusing for a command to skip if it was invoked directly, i.e.,
not as a dependency of a different command.