build: Remove `@wordpress/core-commands` usage
This existing import resulted in the following Vite build error. This
began occurring after upgrading the package. Presumably, this began
occurring due to the addition of the `exports` field to the package in:
https://github.com/WordPress/gutenberg/commit/f13dcfaa60e08a4b0ac77dc105bb8b1eac055264

```
(!) Failed to run dependency scan. Skipping dependency pre-bundling. Error: The following dependencies are imported but could not be resolved:

  @wordpress/core-commands (imported by ./GutenbergKit/src/utils/wordpress-globals.js)

Are they installed?
    at file://./GutenbergKit/node_modules/vite/dist/node/chunks/config.js:31896:33
    at process.processTicksAndRejections (node:internal/process/task_queues:95:5)
    at async file://./GutenbergKit/node_modules/vite/dist/node/chunks/config.js:34104:15

```